requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.9
bcrypt==4.1.1
PyJWT==2.8.0
python-decouple==3.8
